Feat(EREF-59) : Audit Trail/Logs and some fixes (#50)
* added common test library

* refactor(efti-999): add solution for null requestId

* feat(logs) : Added new logs for message reception from gate or platforms

* feat() update endpoint get identifiers result to add request status

* fixed some sonar issues

---------

+++ implementation/registry-of-identifiers/src/main/java/eu/efti/identifiersregistry/repository/IdentifiersRepository.java
@@ -116,7 +116,7 @@ public interface IdentifiersRepository extends JpaRepository<Consignment, Long>,
             predicates.add(cb.and(cb.equal(mainCarriageTransportMovementJoin.get(IS_DANGEROUS_GOODS), request.getDangerousGoodsIndicator())));
         }
         if (StringUtils.isNotBlank(request.getModeCode())) {
-            predicates.add(cb.and(cb.equal(mainCarriageTransportMovementJoin.get(TRANSPORT_MODE), Short.valueOf(request.getModeCode()))));
+            predicates.add(cb.and(cb.equal(mainCarriageTransportMovementJoin.get(TRANSPORT_MODE), request.getModeCode())));
         }
     }
 }
